lab5/src/app/cart.service.ts
import { HttpClient } from '@angular/common/http';
import { Product } from './products';
import { Injectable } from '@angular/core';
/* . . . */
@Injectable({
  providedIn: 'root'
})
export class CartService {
  items: Product[] = [];

  constructor(
    private http: HttpClient
  ) {}

  addToCart(product: Product) {
    this.items.push(product);
  }

  getItems() {
    return this.items;
  }

  clearCart() {
    if (this.items.length > 1){
      window.alert('Your products would be delivered for you soon!');
      this.items = [];
      return this.items
    } else{
      window.alert('Your product would be delivered for you soon!');
      this.items = [];
      return this.items;
    }
  }

  getShippingPrices() {
    return this.http.get<{type: string, price: number}[]>('/assets/shipping.json');
  }
}